reestructuracion de las carpetas en el frontend
