+++ frontend/src/App.tsx
@@ -1,5 +1,5 @@
 import { RouterProvider } from 'react-router-dom';
-import { router } from './routes';
+import { router } from './router';
 
 const App = () => {
   return <RouterProvider router={router} />;
