--- frontend/src/features/home/components/Home.tsx
@@ -1,7 +1,0 @@
-import { Typography } from 'antd';
-
-const { Title } = Typography;
-
-export const Home = () => {
-  return <Title>Home</Title>;
-};
