+++ frontend/src/pages/HomePage.tsx
@@ -1,0 +1,12 @@
+import { Typography } from 'antd';
+
+const { Title } = Typography;
+
+export const HomePage = () => {
+  return (
+    <div style={{ padding: '24px' }}>
+      <Title>Inicio</Title>
+      {}
+    </div>
+  );
+};
